+++ src/main/resources/release_notes.md
@@ -35,7 +35,7 @@
 
 ### Lineup
 
-* Refactoring of lineup assistant (#163)
+* Lineup assistant optimizes behaviour settings (#163)
 * Fix substitution plausibility check (#2224)
 
 ### Special events